fix: remove hacky read/writeQuery, replace with nice local resolvers
This makes everything a lot cleaner, the UI updates itself based on the local state now without hacks, and everything is generally a lot nicer.

This is also introduces a great bug where you can't zoom, and you'll bounce back to the initial zoom. I have no idea what's causing it but intend to fix it in the next commit.