test: use MSR to supply mocks
When we moved from Apollo Client to SWR our previous apporach of
supplying mocks as an executable schema to apollo-link-schema when
testing was invalidated.

@stphnnn pointed out https://mswjs.io/ and it is super-nice! You get to
see your mocks in the network tab without having to run a server, it
plays well with pupeteer, and I
believe we can reuse the same mocks in unit tests.

We have to perform a little shimmy because MSW's graphql support isn't
expecting an executableSchema (for good reasons), maybe that'll change
one day.