feat: adds new Badge component not using Emotion
This is start of experimenting with how we will move away from using
Emotion to style the app. This Badge is made with React and plain old
css. It has been used for when folk remove a toliet they get flagged
that they have successfully removed a loo.

See this issue for more context on the move away from using Emotion for
our styling needs:
https://github.com/public-convenience-ltd/toiletmap/issues/1446