Refactor Drawer component and remove framer-motion dependency
Closes #1654

This commit replaces the `framer-motion` dependency with a custom CSS
transition.

To keep the implementation simple, the `visibility` property is being
toggled instead of `display`, as it still removes the element from the
accessibility tree but has the benefit of being animatable. As the
element has a fixed position, it should not have significantly worse
performance.