Feature: Adds and refactors Header component into design system
Removes "find a toilet" link in main menu as discussed in
Slack it felt unneccessary and just duplicated the home
logo link behaviour.

Aligns all main menu content to left on larger devices as the
menu is now smaller.

Also removes footer from hamburger menu on smaller devices as
a `footer` element in a `nav` felt wrong and probably
unnecessary.

Improves accessibility for hamburger menu on smaller devices by
adding a focus trap and the ability to close using esc for
keyboard users.